Creating database entities.

+++ app/src/main/java/com/jgeniselli/banco/game/common/domain/PlayerRepository.kt
@@ -2,7 +2,7 @@ package com.jgeniselli.banco.game.common.domain
 
 interface PlayerRepository {
     fun findAll(onSuccess: (List<Player>) -> Unit, onError: () -> Unit)
-    fun findById(playerId: Int, onSuccess: (Player?) -> Unit)
+    fun findById(playerId: Long, onSuccess: (Player?) -> Unit)
     fun createPlayer(color: Color): Player
 }
